Remove useless steps and replace the step order.
Ensure the flow works as previously
